% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{cal2}
\alias{cal2}
\title{Range of calculations from the confusion matrix}
\usage{
cal2(x, y, cal)
}
\arguments{
\item{x}{A \code{matrix} that holds the independent variables}

\item{y}{A \code{vector} representing the dependent variable}

\item{cal}{A \code{character} representing the value to be computed. "pre" calculates prevelance, "acc" calculates accuracy, "sen" calculates sensitivity, "spe" calculates "specificity, "fdr" calculates false discovery ratio, "dor" calculates diagnostic odds ratio}
}
\value{
A \code{character} of the computer value
}
\description{
This function calculates various values from the confusion matrix using series of cuttoffs between 0.1-0.9
}
\author{
Manisha Parajuli, Favour Onyido, Festus Attah
}
